requirements.txt
python-telegram-bot==13.15
python-dotenv==1.0.0
pymongo==4.5.0 